Tests/LibWeb/Text/expected/URL/search-params-strip-trailing-spaces-from-opaque-url.txt

URL pathname is 'space    '
URL href is 'data:space    ?test'
true
false
URL pathname is 'space'
URL href is 'data:space'
